requirements.txt
brainflow==5.16.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.58.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pillow==11.2.1
pylsl==1.17.6
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
zipp==3.21.0
